requirements.txt
trimesh
numpy==1.21.2
pyhocon
icecream
opencv_python==4.5.5.62
tqdm
scipy
PyMCubes
lpips
matplotlib
termcolor
tensorboard
pytorch-lightning==1.5.10
pyrender
libigl
mesh2sdf
Cython
wandb